Add TensorBoard to Notebook 02
This commit introduces name_scopes for the convolutional layers, for their
weights and biases, and for the fully-connected layers, so that these
items show up with meaningful names in TensorBoard.

This commit also introduces scalar summaries for the accuracy and cost,
which are calculated every 100th optimization.

Additionally, this commit adds histogram summaries for the weights and biases,
so that how these tensors change over time can be more easily seen.

The default log directory has been set to "logs", which will create
a new directory "logs" in the same directory as the 02_Convolutional_
Neural_Network.ipynb.

To access TensorBoard, from a terminal, type:
`tensorboard --log-dir=/path/to/02_CNN/folder/logs`